Moved tests used in String widgets to their base classes
