+++ collect_app/src/main/java/org/odk/collect/android/widgets/ExStringWidget.java
@@ -97,7 +97,7 @@ public class ExStringWidget extends StringWidget implements BinaryWidget {
     protected static final String DATA_NAME = "value";
 
     private boolean hasExApp = true;
-    Button launchIntentButton;
+    public Button launchIntentButton;
 
     @Inject
     public ActivityAvailability activityAvailability;
