+++ collect_app/src/test/java/org/odk/collect/android/widgets/ExStringWidgetTest.java
@@ -1,18 +1,13 @@
 package org.odk.collect.android.widgets;
 
-import android.view.View;
-
 import androidx.annotation.NonNull;
 
 import net.bytebuddy.utility.RandomString;
 
 import org.javarosa.core.model.data.StringData;
-import org.junit.Test;
 import org.odk.collect.android.formentry.questions.QuestionDetails;
 import org.odk.collect.android.widgets.base.GeneralExStringWidgetTest;
 
-import static junit.framework.TestCase.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.mockito.Mockito.when;
 
 /**
@@ -37,14 +32,5 @@ public class ExStringWidgetTest extends GeneralExStringWidgetTest<ExStringWidget
     public void setUp() throws Exception {
         super.setUp();
         when(formEntryPrompt.getAppearanceHint()).thenReturn("");
-    }
-
-    @Test
-    public void readOnlyTest() {
-        when(formEntryPrompt.isReadOnly()).thenReturn(true);
-
-        assertEquals(View.GONE, getWidget().launchIntentButton.getVisibility());
-        assertEquals(View.VISIBLE, getWidget().answerText.getVisibility());
-        assertFalse(getWidget().answerText.isEnabled());
     }
 }
