+++ collect_app/src/test/java/org/odk/collect/android/widgets/StringWidgetTest.java
@@ -1,19 +1,12 @@
 package org.odk.collect.android.widgets;
 
-import android.view.View;
-
 import androidx.annotation.NonNull;
 
 import net.bytebuddy.utility.RandomString;
 
 import org.javarosa.core.model.data.StringData;
-import org.junit.Test;
 import org.odk.collect.android.formentry.questions.QuestionDetails;
 import org.odk.collect.android.widgets.base.GeneralStringWidgetTest;
-
-import static junit.framework.TestCase.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.mockito.Mockito.when;
 
 /**
  * @author James Knight
@@ -30,13 +23,5 @@ public class StringWidgetTest extends GeneralStringWidgetTest<StringWidget, Stri
     @Override
     public StringData getNextAnswer() {
         return new StringData(RandomString.make());
-    }
-
-    @Test
-    public void readOnlyTest() {
-        when(formEntryPrompt.isReadOnly()).thenReturn(true);
-
-        assertEquals(View.VISIBLE, getWidget().answerText.getVisibility());
-        assertFalse(getWidget().answerText.isEnabled());
     }
 }
