+++ collect_app/src/test/java/org/odk/collect/android/widgets/base/GeneralExStringWidgetTest.java
@@ -1,7 +1,14 @@
 package org.odk.collect.android.widgets.base;
 
+import android.view.View;
+
 import org.javarosa.core.model.data.IAnswerData;
+import org.junit.Test;
 import org.odk.collect.android.widgets.ExStringWidget;
+
+import static junit.framework.TestCase.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.mockito.Mockito.when;
 
 /**
  * @author James Knight
@@ -12,5 +19,14 @@ public abstract class GeneralExStringWidgetTest<W extends ExStringWidget, A exte
     @Override
     public Object createBinaryData(A answerData) {
         return answerData.getDisplayText();
+    }
+
+    @Test
+    public void readOnlyTest() {
+        when(formEntryPrompt.isReadOnly()).thenReturn(true);
+
+        assertEquals(View.GONE, getWidget().launchIntentButton.getVisibility());
+        assertEquals(View.VISIBLE, getWidget().answerText.getVisibility());
+        assertFalse(getWidget().answerText.isEnabled());
     }
 }
